cleanup docker setup
